add board2state

+++ peg_solitaire.py
@@ -1,4 +1,5 @@
 import numpy as np
+from peg_markov import *
 
 def create_board():
 	board = np.ones((7,7))
@@ -7,4 +8,4 @@ def create_board():
 	board[5: , 0:2] = 0
 	board[5: , 5: ] = 0
 	board[3, 3] = 0
-	return board
+	return board.astype(int)
